requirements.txt
Flask_RESTful==0.3.8
rake_nltk==1.0.3
nltk==3.4.5
gensim==3.7.3
SQLAlchemy==1.3.16
Flask_Cors==3.0.8
Flask_SQLAlchemy==2.4.0
flask_marshmallow==0.12.0
Flask==1.1.2
Flask_Migrate==2.5.3
Flask-Script==2.0.6
marshmallow_sqlalchemy==0.23.0
pandas==1.0.3
azure-cognitiveservices-speech==1.12.0
noisereduce==1.1.0